+++ CHANGES.md
@@ -4,7 +4,8 @@
 * Fixed disabled emitters not being exported when saving
 * Window size is initialized to a percentage of the available screen size
 * Added more details to the summary of the Wizard mode
-* Added the Delta Multiplier functionality of the preview 
+* Added the Delta Multiplier functionality of the preview
+* Added the Pixels Per Meter functionality of the preview
 * Minor bugfixes and code cleanup
 ### Version 0.0.2
 * Initial release version